Improve project structure
