+++ newspaper/templates/signup.html
