+++ requirements.txt
@@ -1,6 +1,6 @@
 Flask==2.2.2
-beautifulsoup4==4.11.1
-requests==2.28.1
+Flask-Mail==0.9.1
 Flask-SQLAlchemy==3.0.0
 email-validator==1.3.0
-Flask-Mail==0.9.1
+beautifulsoup4==4.11.1
+requests==2.28.1
