--- src/config.cfg.template
@@ -1,9 +1,0 @@
-SECRET_KEY=""
-SQLALCHEMY_DATABASE_URI="sqlite:///tmn.db"
-MAIL_SERVER="smtp.gmail.com"
-MAIL_PORT=465
-MAIL_USE_TLS=False
-MAIL_USE_SSL=True
-MAIL_USERNAME="username"
-MAIL_PASSWORD="password"
-MAIL_DEFAULT_SENDER=("TheTechMorningNewspaper", "sender")
